refactor: move updateSession to supabase directory
- Extract updateSession function to apps/nextjs/src/supabase/middleware.ts
- Update main middleware.ts to import from supabase directory
- Better code organization following Supabase documentation patterns
